backend/requirements.txt
Django=='5.0.2'
django-filter=='23.5'
djangorestframework=='3.14.0'
Markdown=='3.5.2'
